checked out develop branch, refactored app structure
