+++ .gitignore
@@ -1,7 +1,7 @@
 /.idea
-dao/__pycache__
+app/dao/__pycache__
 helpers/__pycache__
-views/__pycache__
+app/views/__pycache__
 .DS_Store
 /temp
 __pycache__
