+++ app/config.py
@@ -1,6 +1,6 @@
 class Config(object):
     DEBUG = True
     SECRET_HERE = '249y823r9v8238r9u'
-    SQLALCHEMY_DATABASE_URI = 'sqlite:///./movies.db'
+    SQLALCHEMY_DATABASE_URI = 'sqlite:///movies.db'
     SQLALCHEMY_TRACK_MODIFICATIONS = False
