+++ app/container.py
@@ -1,10 +1,10 @@
 from dao.director import DirectorDAO
 from dao.genre import GenreDAO
 from dao.movie import MovieDAO
-from service.director import DirectorService
-from service.genre import GenreService
-from service.movie import MovieService
-from setup_db import db
+from services.director import DirectorService
+from services.genre import GenreService
+from services.movie import MovieService
+from app.setup_db import db
 
 director_dao = DirectorDAO(session=db.session)
 genre_dao = GenreDAO(session=db.session)
