+++ app/dao/director.py
@@ -1,4 +1,4 @@
-from dao.model.director import Director
+from app.dao.models.director import Director
 
 
 class DirectorDAO:
