+++ app/dao/models/director.py
@@ -1,14 +1,14 @@
 from marshmallow import Schema, fields
 
-from setup_db import db
+from app.setup_db import db
 
 
-class Genre(db.Model):
-    __tablename__ = 'genre'
+class Director(db.Model):
+    __tablename__ = 'director'
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(255))
 
 
-class GenreSchema(Schema):
+class DirectorSchema(Schema):
     id = fields.Int()
     name = fields.Str()
