+++ app/dao/models/genre.py
@@ -1,14 +1,14 @@
 from marshmallow import Schema, fields
 
-from setup_db import db
+from app.setup_db import db
 
 
-class Director(db.Model):
-    __tablename__ = 'director'
+class Genre(db.Model):
+    __tablename__ = 'genre'
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(255))
 
 
-class DirectorSchema(Schema):
+class GenreSchema(Schema):
     id = fields.Int()
     name = fields.Str()
