+++ app/services/movie.py
@@ -1,4 +1,4 @@
-from dao.movie import MovieDAO
+from app.dao.movie import MovieDAO
 
 
 class MovieService:
