+++ app/views/directors.py
@@ -1,7 +1,7 @@
 from flask_restx import Resource, Namespace
 
-from dao.model.director import DirectorSchema
-from implemented import director_service
+from app.dao.models.director import DirectorSchema
+from app.container import director_service
 
 director_ns = Namespace('directors')
