+++ app/views/movies.py
@@ -1,8 +1,8 @@
 from flask import request
 from flask_restx import Resource, Namespace
 
-from dao.model.movie import MovieSchema
-from implemented import movie_service
+from app.dao.models.movie import MovieSchema
+from app.container import movie_service
 
 movie_ns = Namespace('movies')
