+++ main.py
@@ -1,11 +1,11 @@
 from flask import Flask
 from flask_restx import Api
 
-from config import Config
-from setup_db import db
-from views.directors import director_ns
-from views.genres import genre_ns
-from views.movies import movie_ns
+from app.config import Config
+from app.setup_db import db
+from app.views.directors import director_ns
+from app.views import genre_ns
+from app.views.movies import movie_ns
 
 
 def create_app(config_object):
